ms3-visionPlante-main/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
minio==7.2.3
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
python-dotenv==1.0.1
pydantic-settings>=2.0.0
torchvision==0.16.2
opencv-python-headless==4.9.0.80
